pyproject.toml
[project]
name = "fu-charsheet"
version = "0.1.0"
description = "Fabula Ultima Charsheet"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "streamlit>=1.47.0",
]
